Move timeline to its own folder and extract Events out of SpellTimeline
